# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API documentation for the golang backend server.",
        "title": "Go Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/selections": {
            "get": {
                "description": "Retrieves selections based on either a document's UUID or a specific selection's UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "selections"
                ],
                "summary": "Get selections by document or selection UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The UUID of the document to retrieve selections for",
                        "name": "documentUUID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The UUID of the specific selection to retrieve",
                        "name": "selectionUUID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of selections",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Selection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, typically due to missing/invalid UUID parameter"
                    },
                    "500": {
                        "description": "Internal server error, "
                    }
                }
            },
            "post": {
                "description": "Creates a new selection associated with a document.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "selections"
                ],
                "summary": "Add a new selection",
                "parameters": [
                    {
                        "description": "Selection creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddNewSelectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful creation, returns the selection UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, typically due to invalid input"
                    },
                    "500": {
                        "description": "Internal server error, "
                    }
                }
            },
            "delete": {
                "description": "Deletes selections based on a specific selection UUID or all selections associated with a document UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "selections"
                ],
                "summary": "Delete selections by selection or document UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The UUID of the specific selection to delete",
                        "name": "selectionUUID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The UUID of the document whose selections are to be deleted",
                        "name": "documentUUID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deletion",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, typically due to missing/invalid UUID parameter"
                    },
                    "500": {
                        "description": "Internal server error, "
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Selection": {
            "type": "object",
            "properties": {
                "documentUUID": {
                    "type": "string"
                },
                "isComplete": {
                    "type": "boolean"
                },
                "selectionBounds": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/models.SelectionBounds"
                        }
                    }
                },
                "selectionUUID": {
                    "type": "string"
                },
                "settings": {
                    "type": "string"
                }
            }
        },
        "models.SelectionBounds": {
            "type": "object",
            "properties": {
                "extract_method": {
                    "type": "string",
                    "example": "None"
                },
                "x1": {
                    "type": "number",
                    "example": 43.122
                },
                "x2": {
                    "type": "number",
                    "example": 13
                },
                "y1": {
                    "type": "number",
                    "example": 52.125
                },
                "y2": {
                    "type": "number",
                    "example": 27.853
                }
            }
        },
...
        "v1.AddNewSelectionRequest": {
            "type": "object",
            "properties": {
                "documentUUID": {
                    "type": "string"
                },
                "isComplete": {
                    "type": "boolean"
                },
                "selectionBounds": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/models.SelectionBounds"
                        }
                    }
                },
                "settings": {
                    "type": "string"
                }
            }
        },
...
    },
...
}